[Tracking bug 26013] Fixing build warnings
[git-p4: depot-paths = "//depot/software/smrtanalysis/bioinformatics/staging/PostPrimary/pbbam/": change = 153842]

+++ include/pbbam/internal/DataSetElement.h
@@ -252,7 +252,7 @@ inline T& DataSetElement::Child(const std::string& label)
 {
     const int i = IndexOf(label);
     if (i >= 0) {
-        assert(i < NumChildren());
+        assert(static_cast<size_t>(i) < NumChildren());
         return Child<T>(i);
     } else {
         AddChild(DataSetElement(label));
